Fix 1.15.0 linker errors: LOCALID symbols, StartNewPokeballCaseUI
- Add 25 missing LOCALID_* defines to map_event_ids.h for BPE custom
  NPCs (rival/Cynthia appearances, bridge Kecleon/Steven, Gerald, etc.)
- Add 3 missing object events to Route120/map.json: BRIDGE_KECLEON,
  BRIDGE_KECLEON_SHADOW, ROUTE120_STEVEN (positions 43-45)
- Restore StartNewPokeballCaseUI to src/script.c (lost in merge)
- Restore BirchCase_GiveMonParameterized to src/script_pokemon_util.c
  with updated type conversions for 1.15.0 (u8->u16 evs/ivs arrays,
  bool8->enum ShinyMode, enum Move* cast)
- Add includes for ui_birch_case.h, task.h, field_weather.h in script.c

Build: clean, 25988552 B ROM (77.45% of 32MB)

+++ src/field_specials.c
@@ -562,7 +562,7 @@ void SpawnLinkPartnerObjectEvent(void)
     };
     u8 myLinkPlayerNumber;
     enum Direction playerFacingDirection;
-    u8 linkSpriteId;
+    u16 linkSpriteId;
     u8 i;
 
     myLinkPlayerNumber = GetMultiplayerId();
